updated LoginWidget in static-pages and edu letters app
