+++ src/applications/education-letters/containers/App.jsx
@@ -1,17 +1,11 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import FormTitle from 'platform/forms-system/src/js/components/FormTitle';
-import { toggleLoginModal as toggleLoginModalAction } from 'platform/site-wide/user-nav/actions';
 import PropTypes from 'prop-types';
 import Layout from '../components/Layout';
 import LoginWidget from '../components/LoginWidget';
 
-const App = ({ toggleLoginModal, user }) => {
-  function toggleLogin(e) {
-    e.preventDefault();
-    toggleLoginModal(true, 'cta-form');
-  }
-
+const App = ({ user }) => {
   function renderUI() {
     if (!user?.login?.currentlyLoggedIn && !user?.login?.hasCheckedKeepAlive) {
       return (
@@ -28,7 +22,7 @@ const App = ({ toggleLoginModal, user }) => {
       window.location.href = '/education/download-letters/letters';
     }
 
-    return <LoginWidget toggleLogin={toggleLogin} />;
+    return <LoginWidget />;
   }
 
   return (
@@ -97,7 +91,6 @@ const App = ({ toggleLoginModal, user }) => {
 };
 
 App.propTypes = {
-  toggleLoginModal: PropTypes.func,
   user: PropTypes.object,
 };
 
@@ -105,11 +98,4 @@ const mapStateToProps = state => ({
   user: state.user || {},
 });
 
-const mapDispatchToProps = dispatch => ({
-  toggleLoginModal: open => dispatch(toggleLoginModalAction(open)),
-});
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(App);
+export default connect(mapStateToProps)(App);
